backend/requirements.txt
fastapi==0.115.14
fastmcp==2.9.2
langchain==0.3.26
langchain_community==0.3.26
langchain_google_vertexai==2.0.26
langchain_qdrant==0.2.0
mcp==1.10.0
passlib==1.7.4
pydantic==2.11.7
python-dotenv==1.1.1
qdrant_client==1.14.3
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.47.1
